+++ frame-web/src/main/resources/businessdb.sql
@@ -28,14 +28,14 @@ create table `nd_exam_info`
     `term`           tinyint(1) comment '学期，1：第一学期，2：第二学期',
     `student_name`   varchar(32) comment '学生姓名',
     `student_number` varchar(32) comment '学生学号',
-    `allscore` int(11) comment '总学分',
-    `getscore` int(11) comment '获得学分',
+    `allscore`       float(5, 2) comment '总学分',
+    `getscore`       float(5, 2) comment '获得学分',
     `gpa`            float(4, 2) default 0.00 comment '学生绩点',
-    `rank` int(11) comment '排名',
+    `rank`           int(11) comment '排名',
     `create_time`    datetime    default current_timestamp comment '创建时间',
     `update_time`    datetime comment '更新时间',
     `delete_time`    datetime comment '删除时间',
-    index create_time_index (create_time desc ),
+    index create_time_index (create_time desc),
     foreign key (student_number) references nd_student (number)
 ) engine = innodb
   default charset = utf8mb4 comment ='学生成绩表';
@@ -51,7 +51,7 @@ create table `nd_exam_course`
     `create_time`  datetime    default current_timestamp comment '创建时间',
     `update_time`  datetime comment '更新时间',
     `delete_time`  datetime comment '删除时间',
-    index create_time_index (create_time desc ),
+    index create_time_index (create_time desc),
     foreign key (exam_info_id) references nd_exam_info (id)
 ) engine = innodb
   default charset = utf8mb4 comment ='学生成绩科目表';
